requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0

# Vector Database
chromadb>=0.5.0

# Web Search
tavily-python>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Optional but useful
pandas>=2.0.0
numpy>=1.24.0